requirements.txt
lightning==2.5.3
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0
PyYAML==6.0.2
scikit_image==0.19.3
torch==2.2.2
torchvision==0.13.0
tensorboard== 2.16.2
tensorboardX==2.6.4